% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampledata.R
\docType{data}
\name{exampledata}
\alias{exampledata}
\title{Example Data from 2 Facebook pages - posts and comments}
\format{
An object of class \code{"data.frame"}
\describe{
\item{level}{node Level, a value between 0 and 2}
\item{id}{own ID in the data set, chronological numbers starting at 1}
\item{parent_id}{ID of the parent node in the data set, from which the data was fetched}
\item{object_id}{own ID,in the given platform format,
here two 15/16 digits numbers for the page and the post "XXX_XXX"}
\item{object_type}{kind of node, possible types: seed, data, offcut, unpacked}
\item{object_key}{keys for the further extraction in Facepager}
\item{query_status}{how far is the query?
was it able to collect data, was there an error message, etc.?}
\item{query_time}{Time of the query via facepager}
\item{query_type}{which Facepager module and which resource were used for the query}
\item{message}{text of the post/comment}
\item{created_time}{time of publication }
\item{like_count}{like count of the comments}
\item{comment_count}{comment count of the comments (replies)}
\item{shares.count}{shares count of the posts}
\item{comments.summary.total_count}{comment count of the posts}
\item{permalink_url}{URL of the posts}
\item{reactions.summary.total_count}{reaction count of the posts}
\item{attachments.data.*.title}{caption of the post, picture etc.}
}
}
\usage{
data(exampledata)
}
\description{
A small dataset generated as an example,
where the unit of analysis is either a post or a comment.
It contains information for 2 posts per page and their comments posted in November 2022
to illustrate the use of the package.
The dataset is exactly in the same state as you export it from Facepager as a csv file.
This means that the columns are named as they were requested from the API.
}
\references{
This data set was created as an example for the facepager package.
}
\keyword{datasets}
