% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facepageR.R
\docType{package}
\name{facepageR-package}
\alias{facepageR}
\alias{facepageR-package}
\title{facepageR: Analyze Facepager Data with R}
\description{
To learn more about the facepageR package, start with the vignettes:
\code{browseVignettes(package = "facepageR")}
They are divided into load and prepare data, basics and utilities.
They may help you to understand the package better and show you how to use it best.
}
\seealso{
Useful links: tbc
}
\author{
Jakob Jünger
}
