% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facepager_function.R
\name{fp_childcount}
\alias{fp_childcount}
\title{Die Anzahl unterschiedlicher Kinder je Object ID auszählen (ohne Duplikate)}
\usage{
fp_childcount(.data, .parents = NA, .silent = FALSE, level = 0)
}
\description{
Die Anzahl unterschiedlicher Kinder je Object ID auszählen (ohne Duplikate)
}
