% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_children.R
\name{fp_children}
\alias{fp_children}
\title{get the collected data per Object ID}
\usage{
fp_children(
  .data,
  col_target = NA,
  col_paging = NA,
  col_created = NA,
  timestamp = F,
  level = 0
)
}
\arguments{
\item{.data}{the loaded data from Facepager}

\item{timestamp}{on Twitter, the timestamp must be parsed (timestamp == "twitter"),
on Instagram, the timestamp must first be converted to a date (timestamp == TRUE)}

\item{col_target=NA}{coverage per object ID}

\item{col_paging=NA}{offcuts and last cursor per object ID}

\item{col_created=NA}{first and last date per object ID}

\item{level=0}{the node level}
}
\value{
a tibble with ?
}
\description{
get the collected data per Object ID
}
