% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facepageR_function.R
\name{fp_comparefiles}
\alias{fp_comparefiles}
\title{Function to synchronize files between facepager and folder}
\usage{
fp_comparefiles(data, folder = "../profiledata/")
}
\description{
Function to synchronize files between facepager and folder
}
