% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facepager_function.R
\name{fp_created}
\alias{fp_created}
\title{Erstes und letztes Datum je Object ID}
\usage{
fp_created(.data, col_created, timestamp = FALSE, .parents = NA, level = 0)
}
\description{
Erstes und letztes Datum je Object ID
}
