% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_duplicates.R
\name{fp_duplicates}
\alias{fp_duplicates}
\title{Evaluate duplicates on different node levels}
\usage{
fp_duplicates(data, level = 0)
}
\arguments{
\item{data}{the loaded data from Facepager}

\item{level}{the node level}
}
\value{
a list of duplicates at parent level,
duplicates within the seeds, duplicates at all childnodes
}
\description{
With this function duplicates will be extracted on three different levels:
\enumerate{
\item at parent level,
\item at children level with same parent/seed
\item at children level independent of parent
}
}
