% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_extracttags.R
\name{fp_extracttags}
\alias{fp_extracttags}
\title{get tags (urls, mentions, hashtags)}
\usage{
fp_extracttags(data, col_text, prefix = NA)
}
\arguments{
\item{data}{the loaded data from Facepager}

\item{col_text}{the column from which the tags should be extracted}

\item{prefix=NA}{when transformatting, the values are not prefixed with a character}
}
\value{

}
\description{
get tags (urls, mentions, hashtags)
}
