% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractdata.R
\name{fp_getparents}
\alias{fp_getparents}
\title{Get the Object IDs  and number of parent nodes}
\usage{
fp_getparents(.data, level = 0)
}
\arguments{
\item{data}{the loaded data from Facepager}

\item{level=0}{the node level}
}
\value{
a tibble with 3 columns: parent_id, parent_object, parent_no
}
\description{
\code{fp_getparents()} can be used to get details about the parents.
}
\examples{
fp_getparents(data)
}
