% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_getthreads.R
\name{fp_getthreads}
\alias{fp_getthreads}
\title{Transfer tree_thread (ID of root node) to all children of root nodes}
\usage{
fp_getthreads(data, col_id, col_parent, col_order = NULL)
}
\arguments{
\item{data}{All nodes (e.g. posts and comments)}

\item{col_id}{The column holding IDs of the nodes}

\item{col_parent}{The column holding IDs of the parent nodes}

\item{col_order}{The column used for ordering the nodes inside of a level}
}
\description{
Transfer tree_thread (ID of root node) to all children of root nodes
}
