% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_gettree.R
\name{fp_gettree}
\alias{fp_gettree}
\title{Adds left and right values to the dataframe (modified preorder tree traversal)
See https://www.sitepoint.com/hierarchical-data-database-3/}
\usage{
fp_gettree(data)
}
\arguments{
\item{data}{dataframe with tree_level, tree_thread, tree_id, tree_parent, created}
}
\description{
Adds left and right values to the dataframe (modified preorder tree traversal)
See https://www.sitepoint.com/hierarchical-data-database-3/
}
