% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_load_databases.R
\name{fp_load_database}
\alias{fp_load_database}
\title{Load a fp database}
\usage{
fp_load_database(dbname, shard = NA)
}
\arguments{
\item{shard}{prefix ids to load multiple databases}

\item{filename}{name of the fp database you want to load (file extension db)}
}
\value{
a data frame containing the data of the fp database
}
\description{
Load a fp database
}
