% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_load_databases.R
\name{fp_load_databases}
\alias{fp_load_databases}
\title{Load multiple fp databases}
\usage{
fp_load_databases(filenames, fields = NULL)
}
\arguments{
\item{filenames}{names of the fp databases you want to load}

\item{fields}{select your collected fields you want to load
(you can only load fields you have collected via facepager)
e.g. objectid, object_type, query_status, query_type, query_time etc.}
}
\value{
A data frame containing the data of multiple fp databases
}
\description{
Load multiple fp databases
}
