% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_loaddatabase.R
\name{fp_loaddatabase}
\alias{fp_loaddatabase}
\title{Load a complete fp database}
\usage{
fp_loaddatabase(dbname, shard = NA)
}
\arguments{
\item{shard}{prefix ids to load multiple databases}

\item{filename}{name of the fp database you want to load (file extension db)}
}
\value{
a data frame containing the data of the fp database
}
\description{
Load a complete fp database
}
