% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facepager_function.R
\name{fp_mentionids}
\alias{fp_mentionids}
\title{#' Extracts the first mention in the text and joins the
#' id of the mentioned node by author name
fp_mentionids <- function(.data,id,parent,no,author,text) {}
\usage{
fp_mentionids(.data, id, parent, no, author, text)
}
\description{
id <- enquo(id)
parent <- enquo(parent)
no <- enquo(no)
author <- enquo(author)
text <- enquo(text)
}
\details{

}
