% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facepageR_function.R
\name{fp_nodesperseed}
\alias{fp_nodesperseed}
\title{Function to}
\usage{
fp_nodesperseed(
  .data,
  col_created,
  daterange = NULL,
  prefix = NA,
  col_seed = NULL
)
}
\description{
Function to
}
