% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facepager_function.R
\name{fp_paging}
\alias{fp_paging}
\title{Anzahl der offcuts und letzter cursor je object ID}
\usage{
fp_paging(.data, col_paging, .parents = NA, level = 0)
}
\description{
Anzahl der offcuts und letzter cursor je object ID
}
