% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_read_csv2.R
\name{fp_read_csv2}
\alias{fp_read_csv2}
\title{Load a CSV file}
\usage{
fp_read_csv2(filename)
}
\arguments{
\item{filename}{name of the CSV file you want to load}

\item{na="None"}{define missing values as string 'None'}
}
\value{
a tibble containing the data of the csv file
}
\description{
Load a CSV file
}
