% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facepageR_function.R
\name{fp_status}
\alias{fp_status}
\title{Function to get number of levels, querytypes, querystatus and objecttypes}
\usage{
fp_status(data)
}
\description{
Function to get number of levels, querytypes, querystatus and objecttypes
}
