% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_tree.R
\name{fp_tree}
\alias{fp_tree}
\title{Creates IDs for threads, adds level and order data}
\usage{
fp_tree(.data, id, parent, no)
}
\arguments{
\item{.data}{The dataframe containing hierarchical data}

\item{id}{The ID of the node}

\item{parent}{The ID of the parent node}

\item{no}{The number of the node inside of parents}
}
\description{
Creates IDs for threads, adds level and order data
}
