% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facepager_function.R
\name{fp_treeids}
\alias{fp_treeids}
\title{Create IDs to avoid collision between levels}
\usage{
fp_treeids(.data, id, parent, type, levels)
}
\arguments{
\item{id}{Column holding the ID of the node}

\item{parent}{Column holding the parent ID of the node}

\item{type}{Column holding the type of the node. Every level needs a different type}

\item{levels}{The hierarchy of the types, e.g. c("forum","post","comment","reply")}
}
\description{
Create IDs to avoid collision between levels
}
