% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facepager_function.R
\name{rebuild_tree}
\alias{rebuild_tree}
\title{Adds left and right values to the dataframe (modified preorder tree traversal)
See https://www.sitepoint.com/hierarchical-data-database-3/}
\usage{
rebuild_tree(data, col_id, col_parent, col_order = NULL)
}
\arguments{
\item{col_id}{Column holding the ID of the node}

\item{col_parent}{Column holding the parent ID of the node}

\item{col_order}{Column holding the order of nodes inside of a parent (e.g. a date). Default is col_id}
}
\description{
Adds left and right values to the dataframe (modified preorder tree traversal)
See https://www.sitepoint.com/hierarchical-data-database-3/
}
