% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facepageR_function.R
\name{rebuild_tree_mptt}
\alias{rebuild_tree_mptt}
\title{Recursively called from rebuild_tree
See https://www.sitepoint.com/hierarchical-data-database-3/}
\usage{
rebuild_tree_mptt(data, parentvalue, left = 1)
}
\arguments{
\item{data}{A dataframe with columns .tree_id,.tree_parent and .tree_order}

\item{parentvalue}{The .tree_id value of the root node}

\item{left}{The left value of the root node}
}
\description{
Recursively called from rebuild_tree
See https://www.sitepoint.com/hierarchical-data-database-3/
}
