% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{tree_disambiguate_ids}
\alias{tree_disambiguate_ids}
\title{Create tree IDs - disambiguate IDs of different node types to avoid collision between IDs}
\usage{
tree_disambiguate_ids(.data, id, parent, type, levels)
}
\arguments{
\item{id}{Column holding the ID of the node}

\item{parent}{Column holding the parent ID of the node}

\item{type}{Column holding the type of the node. Every level needs a different type, e.g. "page", "post", "comment" or "reply".}

\item{levels}{The hierarchy of the types, e.g. c("page","post","comment","reply")}
}
\value{

}
\description{
TODO: test and refine
}
\details{
Work in progress
}
