% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_childcount.R
\name{fp_childcount}
\alias{fp_childcount}
\title{Count the number of different children per Object ID (without duplicates).}
\usage{
fp_childcount(.data, .parents = NA, .silent = FALSE, level = 0)
}
\arguments{
\item{.data}{the loaded data from Facepager}

\item{.parents=NA}{ensures the usage of the parent nodes}

\item{.silent=FALSE}{standard that error messages are displayed}

\item{level=0}{the node level}
}
\value{
a tibble with 4 columns: parent_objectid, children, parent_no, file
}
\description{
Count the number of different children per Object ID (without duplicates).
}
