% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facepageR_function.R
\name{fp_created}
\alias{fp_created}
\title{Function to get the first and the last date per object ID}
\usage{
fp_created(.data, col_created, timestamp = FALSE, .parents = NA, level = 0)
}
\description{
Function to get the first and the last date per object ID
}
