% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractdata.R
\name{fp_created}
\alias{fp_created}
\title{helper function for fp_children()}
\usage{
fp_created(.data, col_created, timestamp = FALSE, .parents = NA, level = 0)
}
\description{
helper function for fp_children()
}
