% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractdata.R
\name{fp_extracttags}
\alias{fp_extracttags}
\title{Get tags included in the nodes}
\usage{
fp_extracttags(data, col_text, prefix = NA)
}
\arguments{
\item{data}{the loaded data from Facepager}

\item{col_text}{the column from which the tags should be extracted}

\item{prefix=NA}{when transformatting, the values are not prefixed with a character}
}
\value{
Columns for the different tags: tags, urls, mentions, count_tags, count_urls, count_mentions, msg_length_chars, msg_length_words
}
\description{
\code{fp_extracttags} can be used to extract all the tags like URLs, mentions or hashtags from the nodes
}
\examples{
fp_extracttags(data, text)
}
