% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facepageR_function.R
\name{fp_getparents}
\alias{fp_getparents}
\title{Function to get Object ID and number of parent nodes}
\usage{
fp_getparents(data, level = 0)
}
\description{
Function to get Object ID and number of parent nodes
}
