% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_getparents.R
\name{fp_getparents}
\alias{fp_getparents}
\title{get the Object IDs  and number of parent nodes}
\usage{
fp_getparents(.data, level = 0)
}
\arguments{
\item{data}{the loaded data from Facepager}

\item{level=0}{the node level}
}
\value{
a tibble with 3 columns: parent_id, parent_object, parent_no
}
\description{
get the Object IDs  and number of parent nodes
}
