% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddata.R
\name{fp_load_database}
\alias{fp_load_database}
\title{Load all nodes from a Facepager database}
\usage{
fp_load_database(
  filename,
  fields = "*",
  rename = T,
  shard = NA,
  .progress = NULL
)
}
\arguments{
\item{filename}{Name of the Facepager database, e.g. "post.db"}

\item{fields}{A character vector of fields to load or "*" to load all fields.
Possible candidates are:
objectid, objecttype, querystatus, querytype, querytime, queryparamse
id, parent_id, level, childcount, response.
The response field contains the data in JSON format.}

\item{rename}{Rename the columns to match the names in Facepager's CSV files}

\item{shard}{When loading multiple databases, give the ids a prefix. This way, the same numerical IDs from different databases don't come into conflict.}

\item{.progress}{progress bar displays estimated time remaining#'}
}
\value{
A data frame containing the data of the selected fields of the fp database.
}
\description{
Load all nodes from a Facepager database
}
\examples{
data_db <- fp_load_database(system.file("extdata", "example.db", package = "facepager"))
}
