% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_databases.R
\name{fp_load_databases}
\alias{fp_load_databases}
\title{Load multiple Facepager databases}
\usage{
fp_load_databases(filenames, fields = "*", rename = TRUE)
}
\arguments{
\item{filenames}{A character vector containing the names of the databases}

\item{fields}{select your collected fields you want to load
(you can only load fields you have collected via facepager)
e.g. objectid, object_type, query_status, query_type, query_time etc.}
}
\value{
A data frame containing the data of multiple fp databases
}
\description{
Load multiple Facepager databases
}
