% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facepageR_function.R
\name{fp_load_nodes}
\alias{fp_load_nodes}
\title{Function to load data from the database}
\usage{
fp_load_nodes(dbname, fields = "*", .progress = NULL)
}
\description{
Function to load data from the database
}
