% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_load_databases.R
\name{fp_load_nodes}
\alias{fp_load_nodes}
\title{Load data from a fp database}
\usage{
fp_load_nodes(dbname, fields = "*", .progress = NULL, shard = NA)
}
\arguments{
\item{dbname}{name of the fp database you want to load data from (file extension db)}

\item{fields}{select your collected fields you want to load
(you can only load fields you have collected via facepager)
e.g. object_id, object_type, query_status, query_type, query_time etc.}

\item{.progress}{progress bar displays estimated time remaining}

\item{shard}{prefix ids to load multiple databases}
}
\value{
A data frame containing the data of the selected fields of the fp database.
}
\description{
Load data from a fp database
}
