% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facepageR_function.R
\name{fp_mentionids}
\alias{fp_mentionids}
\title{Extracts IDs of mentioned comments by comparing all
author names with the comment text
in all subthreads having the same tree_parent.}
\usage{
fp_mentionids(.data, id, parent, no, author, text)
}
\description{
Extracts IDs of mentioned comments by comparing all
author names with the comment text
in all subthreads having the same tree_parent.
}
