% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facepageR_function.R
\name{fp_paging}
\alias{fp_paging}
\title{Function to get the number of offcuts and last cursor per object ID}
\usage{
fp_paging(.data, col_paging, .parents = NA, level = 0)
}
\description{
Function to get the number of offcuts and last cursor per object ID
}
