% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddata.R
\name{fp_parse_response_data}
\alias{fp_parse_response_data}
\alias{fp_from_ndjson}
\alias{fp_get_response_value}
\title{Convert JSON response data to columns of a data frame}
\usage{
fp_parse_response_data(nodes, plain = F, prefix = "response.")

fp_from_ndjson(data)

fp_get_response_value(nodes, col, .split = TRUE, .progress = NULL)
}
\arguments{
\item{nodes}{the fields/columns originally collected by Facepager,
which are all together in the response column,
can be for example "created_at", "text"/ "message", "favorite_count"/ "like_count" etc.}

\item{plain}{Only keep atomic columns, remove data frames and lists}

\item{prefix}{name of the columns}
}
\value{
A data frame containing the data of the Facepager database
with all response data as own columns
}
\description{
As you can see in the loaded data,
most of your data is in JSON format in the response column.
Therefore you need to convert the JSON data in the response to columns of a data frame.
}
\examples{
data_db <- fp_load_database(system.file("extdata", "example.db", package = "facepager"))
data_db <- fp_parse_response_data(data_db)
}
